background_jobs: Move job structs into the corresponding modules
