+++ src/worker/mod.rs
@@ -12,9 +12,10 @@ mod readmes;
 mod update_downloads;
 
 pub(crate) use daily_db_maintenance::perform_daily_db_maintenance;
-pub(crate) use dump_db::perform_dump_db;
+pub(crate) use dump_db::{perform_dump_db, DumpDbJob};
 pub(crate) use git::{
     perform_index_squash, perform_normalize_index, sync_to_git_index, sync_to_sparse_index,
+    NormalizeIndexJob, SyncToIndexJob,
 };
-pub(crate) use readmes::perform_render_and_upload_readme;
+pub(crate) use readmes::{perform_render_and_upload_readme, RenderAndUploadReadmeJob};
 pub(crate) use update_downloads::perform_update_downloads;
